+++ fuse/src/fs.rs
@@ -431,7 +431,7 @@ impl FuserFilesystem for BtrfsFuse {
                     reply.error(Errno::ENODATA);
                     #[cfg(not(target_os = "linux"))]
                     reply.error(Errno::ENOENT);
-                },
+                }
                 Err(e) => {
                     log::warn!(
                         "getxattr ino={fuse_ino} name={}: {e}",